[DEV-866] Webhooks: align retry tests and Redis migrations
