+++ cvat/apps/webhooks/tests/test_signals.py
@@ -6,8 +6,7 @@ from http import HTTPStatus
 from unittest.mock import MagicMock, patch
 
 from django.contrib.auth.models import User
-from django.test import TestCase, override_settings
-from rq import Retry
+from django.test import TestCase
 
 from cvat.apps.engine.models import Project
 from cvat.apps.webhooks.exceptions import WebhookDeliveryError
@@ -17,7 +16,6 @@ from cvat.apps.webhooks.models import (
     WebhookDelivery,
     WebhookTypeChoice,
 )
-from cvat.apps.webhooks.dispatch import add_to_queue
 from cvat.apps.webhooks.tasks import send_webhook
 
 
@@ -41,39 +39,39 @@ def _payload() -> dict:
 
 class TestSendWebhook(TestCase):
     @classmethod
-    def setUpTestData(cls):
+    def setUpTestData(cls) -> None:
         cls.webhook = _make_webhook()
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_raises_on_5xx(self, perform):
+    def test_raises_on_5xx(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.INTERNAL_SERVER_ERROR, "")
 
         with self.assertRaises(WebhookDeliveryError):
             send_webhook(self.webhook.id, _payload())
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_raises_on_502_bad_gateway(self, perform):
+    def test_raises_on_502_bad_gateway(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.BAD_GATEWAY, "")
 
         with self.assertRaises(WebhookDeliveryError):
             send_webhook(self.webhook.id, _payload())
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_raises_on_408_request_timeout(self, perform):
+    def test_raises_on_408_request_timeout(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.REQUEST_TIMEOUT, "")
 
         with self.assertRaises(WebhookDeliveryError):
             send_webhook(self.webhook.id, _payload())
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_raises_on_429_too_many_requests(self, perform):
+    def test_raises_on_429_too_many_requests(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.TOO_MANY_REQUESTS, "")
 
         with self.assertRaises(WebhookDeliveryError):
             send_webhook(self.webhook.id, _payload())
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_does_not_raise_on_2xx(self, perform):
+    def test_does_not_raise_on_2xx(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.OK, '{"echo":true}')
 
         delivery = send_webhook(self.webhook.id, _payload())
@@ -82,7 +80,7 @@ class TestSendWebhook(TestCase):
         assert delivery.status_code == HTTPStatus.OK
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_does_not_raise_on_4xx_other_than_408_429(self, perform):
+    def test_does_not_raise_on_4xx_other_than_408_429(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.BAD_REQUEST, "")
 
         delivery = send_webhook(self.webhook.id, _payload())
@@ -91,7 +89,7 @@ class TestSendWebhook(TestCase):
         assert delivery.status_code == HTTPStatus.BAD_REQUEST
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_5xx_records_delivery_before_raising(self, perform):
+    def test_5xx_records_delivery_before_raising(self, perform: MagicMock) -> None:
         perform.return_value = (HTTPStatus.INTERNAL_SERVER_ERROR, "boom")
 
         with self.assertRaises(WebhookDeliveryError):
@@ -103,7 +101,9 @@ class TestSendWebhook(TestCase):
         assert delivery.event == "update:project"
 
     @patch("cvat.apps.webhooks.tasks._perform_webhook_request")
-    def test_inactive_webhook_returns_none_without_request_or_delivery(self, perform):
+    def test_inactive_webhook_returns_none_without_request_or_delivery(
+        self, perform: MagicMock
+    ) -> None:
         Webhook.objects.filter(pk=self.webhook.id).update(is_active=False)
 
         result = send_webhook(self.webhook.id, _payload())
@@ -111,39 +111,3 @@ class TestSendWebhook(TestCase):
         assert result is None
         perform.assert_not_called()
         assert not WebhookDelivery.objects.filter(webhook=self.webhook).exists()
-
-
-class TestAddToQueue(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        cls.webhook = _make_webhook()
-
-    @override_settings(SEND_WEBHOOK_TASK_RETRIES=[1, 2, 3])
-    @patch("cvat.apps.webhooks.dispatch.django_rq.get_queue")
-    def test_passes_send_webhook_with_retry_intervals(self, get_queue):
-        queue = MagicMock()
-        get_queue.return_value = queue
-
-        add_to_queue(self.webhook, _payload())
-
-        queue.enqueue_call.assert_called_once()
-        kwargs = queue.enqueue_call.call_args.kwargs
-
-        assert kwargs["func"] is send_webhook
-        assert kwargs["args"] == (self.webhook.id, _payload(), False)
-
-        retry: Retry = kwargs["retry"]
-        assert isinstance(retry, Retry)
-        assert retry.max == 3
-        assert retry.intervals == [1, 2, 3]
-
-    @override_settings(SEND_WEBHOOK_TASK_RETRIES=[5, 300, 1800])
-    @patch("cvat.apps.webhooks.dispatch.django_rq.get_queue")
-    def test_redelivery_flag_propagates_to_send_webhook_args(self, get_queue):
-        queue = MagicMock()
-        get_queue.return_value = queue
-
-        add_to_queue(self.webhook, _payload(), redelivery=True)
-
-        kwargs = queue.enqueue_call.call_args.kwargs
-        assert kwargs["args"] == (self.webhook.id, _payload(), True)
